fix(cmpapi): fixes missing tcfPolicyVersion on TCData obj when gdprApplies is false
Policy version currently is only provided by the TC string that is passed to CmpApi.update(). When
gdprApplies is false, that value should be null, however, when its null, there is no signal to
CmpApi what version is supported. The JS API spec requires that the CMP provide a value for this
field, _EVEN WHEN_ GDPR does not apply. The current behavior is that this field is undefined, which
violates that requirement. This change addresses that.